% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{CDF.KNN.2d}
\alias{CDF.KNN.2d}
\title{This function computes M2:kNN-CDF distance in Definition 4 (sparsity measure)}
\usage{
CDF.KNN.2d(sn, k = NULL, cdf_x1 = NULL, cdf_x2 = NULL)
}
\arguments{
\item{k}{an integer constant with value 1 to n representing the number of neighbors to consider in the summation}

\item{cdf_x1}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{cdf_x2}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{data}{sample of dimension n x d, with d = 2}
}
\value{
the M2:kNN-CDF distance for each point in the sample
}
\description{
This function computes M2:kNN-CDF distance in Definition 4 (sparsity measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
CDF.KNN.2d(sn = draws_2d)
}
