% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{CDF.NPCop.2d}
\alias{CDF.NPCop.2d}
\title{This function computes M3-NPCop:e-CDF (concentration measure)}
\usage{
CDF.NPCop.2d(
  sn,
  eps = NULL,
  cdf_x1 = NULL,
  cdf_x2 = NULL,
  fit_cop = NULL,
  method = "TLL2nn"
)
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{eps}{a bidimensional hyperparameter with positive elements, representing the half-length of the rectangular neighborhood}

\item{cdf_x1}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{cdf_x2}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{fit_cop}{a fitted copula model. By default it is set to NULL and is estimated with kde}

\item{method}{method used for estimating the copula with kde; default is "TLL2nn". See kdecopula package for more details: https://cran.r-project.org/web/packages/kdecopula/vignettes/kdecopula.pdf}
}
\value{
the M3-NPCop:e-CDF distance for each point in the sample
}
\description{
This function computes M3-NPCop:e-CDF (concentration measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
CDF.NPCop.2d(sn = draws_2d)
}
