% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{CDF.PCop.2d}
\alias{CDF.PCop.2d}
\title{This function computes M3-PCop:e-CDF (concentration measure)}
\usage{
CDF.PCop.2d(sn, margin_family = c("norm", "norm"), eps = NULL)
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{margin_family}{a two-dimensional vector specifying the two marginal families in the form 'norm', 't', or 'mixnorm'. No other specifications are allowed for the marginals.}

\item{eps}{a bidimensional hyperparameter with positive elements, representing the half-length of the rectangular neighborhood}
}
\value{
the density estimate with M0-PCop:DE distance for each point in the sample
}
\description{
This function computes M3-PCop:e-CDF (concentration measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
CDF.PCop.2d(sn = draws_2d)
}
