% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{DE.NPCop.2d}
\alias{DE.NPCop.2d}
\title{This function computes M0-NPCop:DE (concentration measure)}
\usage{
DE.NPCop.2d(sn, method = "TLL2nn")
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{method}{method used for estimating the copula with kde; default is "TLL2nn". See kdecopula package for more details: https://cran.r-project.org/web/packages/kdecopula/vignettes/kdecopula.pdf}
}
\value{
the density estimate with M0-NPCop:DE distance for each point in the sample
}
\description{
This function computes M0-NPCop:DE (concentration measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
DE.NPCop.2d(sn = draws_2d)
}
