% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainHDR.R
\name{HDR.2d}
\alias{HDR.2d}
\title{This function computes an HDR for an observed sample, given a certain coverage probability}
\usage{
HDR.2d(
  sn,
  measure = "KDE",
  coverage_prob = 0.95,
  build_plot = T,
  k = NULL,
  H = NULL,
  eps = NULL,
  method = "TLL2nn",
  margin_family = c("norm", "norm"),
  cdf_x1 = NULL,
  cdf_x2 = NULL,
  fit_cop = NULL,
  xlab = expression(X[1]),
  ylab = expression(X[2]),
  cex = 0.5,
  col = "darkmagenta",
  pch = 4,
  ...
)
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{measure}{the measure to be used for computing the HDR. Should be one of the following: "KDE", "DE.NPCop", "DE.PCop", "KNN", "CDF.KNN", "CDF.e", "CDF.NPCop", "CDF.PCop".}

\item{coverage_prob}{the coverage probability in (0,1). By default, this is 0.95}

\item{build_plot}{logical, stating whether to plot the HDR classification or not. By default, this is T}

\item{k}{an integer constant with value 1 to n representing the number of neighbors to consider. By default it is = NULL and the rule of thumb is used.}

\item{H}{the bandwidth hyperparameter, in the form of a 2 x 2 matrix. By default it is = NULL and the optimal bandwidth of Chacón, J.E. et al. (2011) is used.}

\item{eps}{a bidimensional hyperparameter with positive elements, representing the half-length of the rectangular neighborhood. By default it is = NULL and the rule of thumb is used.}

\item{method}{method used for estimating the copula with kde; default is "TLL2nn" (see kdecopula package for more details).}

\item{margin_family}{a two-dimensional vector specifying the two marginal families in the form 'norm', 't', or 'mixnorm'. No other specifications are allowed for the marginals. Default is c('norm', 'norm').}

\item{cdf_x1}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{cdf_x2}{an estimate of the cdf of the first marginal. By default it is set to NULL and the empirical cdf is taken}

\item{fit_cop}{a fitted copula model. By default it is set to NULL and it is estimated with kde.}

\item{xlab}{name of the x axis}

\item{ylab}{name of the y axis}

\item{cex}{points dimension}

\item{col}{points color}

\item{pch}{points type}

\item{...}{}
}
\value{
in_out: a vector of 0/1 indicating which data point is inside (1) or outside (0) the HDR; HDR_in: the set of data points from sn classified as HDR data points; HDR_out: the set of data points from sn classified as non-HDR data points
}
\description{
This function computes an HDR for an observed sample, given a certain coverage probability
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
# Estimate the HDR with KDE
HDR.2d(sn = draws_2d, measure = "KDE", coverage_prob = 0.95, build_plot = T)
}
