% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{KDE.2d}
\alias{KDE.2d}
\title{This measure represents M0:KDE (concentration measure)}
\usage{
KDE.2d(sn, H = NULL)
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{H}{the bandwidth hyperparameter, in the form of a 2 x 2 matrix}
}
\value{
the density estimate with M0:KDE distance for each point in the sample
}
\description{
This measure represents M0:KDE (concentration measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
KDE.2d(sn = draws_2d)
}
