% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{KNN.2d}
\alias{KNN.2d}
\title{This function computes M1:kNN-Eucl in Definition 3 (sparsity measure)}
\usage{
KNN.2d(sn, k = NULL)
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{k}{an integer constant with value 1 to n representing the number of neighbors to consider in the summation; by default the rule of thumb is used}
}
\value{
the M1:kNN-Eucl distance for each point in the sample
}
\description{
This function computes M1:kNN-Eucl in Definition 3 (sparsity measure)
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
KNN.2d(sn = draws_2d)
}
