% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_HDR_measures.R
\name{Pfit_margins_cop}
\alias{Pfit_margins_cop}
\title{This function fits a parametric model for each of the two marginals and for the copula model}
\usage{
Pfit_margins_cop(sn, margin_family = c("norm", "norm"))
}
\arguments{
\item{sn}{data sample of dimension n x d, with d = 2}

\item{margin_family}{a two-dimensional vector specifying the two marginal families in the form 'norm', 't', or 'mixnorm'. No other specifications are allowed for the marginals.}
}
\value{
margin_family: a two-dimensional vector specifying the two marginal families; margin_params: a two-dimensional list specifying the estimated parameters of each of the two marginal families; copula_model: the estimated copula model
}
\description{
This function fits a parametric model for each of the two marginals and for the copula model
}
\examples{
# Generate some bivariate data
R = matrix(c(1, 0.5, 0.5, 1), nrow = 2, ncol = 2, byrow = TRUE)
draws_2d = MASS::mvrnorm(n = 2000, mu = c(0, 0), Sigma = R)
Pfit_margins_cop(sn = draws_2d)
}
